+++ lib/xcodeproj/project/project_helper.rb
@@ -157,6 +157,15 @@ module Xcodeproj
 
         cl.build_configurations << release_conf
         cl.build_configurations << debug_conf
+
+        project.build_configurations.each do |configuration|
+          next if cl.build_configurations.map(&:name).include?(configuration.name)
+
+          new_config = project.new(XCBuildConfiguration)
+          new_config.name = configuration.name
+          cl.build_configurations << new_config
+        end
+
         cl
       end